LayoutTests/http/tests/security/frameNavigation/sandbox-DENIED-top-navigation-with-user-gesture-expected.txt

CONSOLE MESSAGE: line 32: Unsafe JavaScript attempt to initiate navigation for frame with URL 'http://127.0.0.1:8000/security/frameNavigation/sandbox-DENIED-top-navigation-with-user-gesture.html' from frame with URL 'http://127.0.0.1:8000/security/frameNavigation/resources/iframe-that-performs-parent-navigation-with-user-activation.html'. The frame attempting navigation of the top-level window is sandboxed, but the 'allow-top-navigation-by-user-activation' flag is not set and navigation is not triggered by user activation.

CONSOLE MESSAGE: line 22: cannot navigate
This tests that a sandboxed iframe without 'allow-top-navigation-by-user-activation' cannot navigate the top level page, even when navigation is trigged by a user gesture.

127.0.0.1
